move getaddrinfo wrapper into util.h
